Implemented the view permissions class.
